Move functions to external files
